ai-service/requirements.txt
fastapi
uvicorn
python-multipart
openai
faiss-cpu
numpy
pydantic
openai-whisper
torch
ffmpeg-python
tiktoken
sentence-transformers
PyMuPDF